+++ lib/screens/admin/scanner.dart
@@ -1,15 +1,18 @@
 import 'dart:collection';
 import 'dart:io';
+
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:camera/camera.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:deca_app/screens/admin/templates.dart';
 import 'package:deca_app/utility/InheritedInfo.dart';
 import 'package:deca_app/utility/notifiers.dart';
+import 'package:deca_app/utility/transition.dart';
 import 'package:firebase_ml_vision/firebase_ml_vision.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
+
+import 'events.dart';
 import 'finderscreen.dart';
 
 class Scanner extends StatefulWidget {
@@ -38,25 +41,22 @@ class _ScannerState extends State<Scanner> {
         .collection("Users")
         .document(userUniqueID)
         .get();
+
     gpContainer.setUserData(userSnapshot.data);
 
-    if (gpContainer.eventMetadata['enter_type'] == 'ME') {
-      gpContainer.setIsManualEnter(true);
-      return "Ok";
-    } else {
-      gpContainer.updateGP(userUniqueID);
-      String firstName = gpContainer.userData['first_name'];
-      _scaffoldKey.currentState.showSnackBar(SnackBar(
+    gpContainer.updateGP(userUniqueID);
+    String firstName = gpContainer.userData['first_name'];
+    _scaffoldKey.currentState.showSnackBar(SnackBar(
         backgroundColor: Color.fromRGBO(46, 204, 113, 1),
         content: Text("Scanned " + firstName),
-        duration: Duration(seconds: 1),
-      ));
-      return "Ok";
-    }
+        duration: Duration(milliseconds: 500)));
+    return "Ok";
   }
 
   void runStream() {
     _scannedUids = new HashSet();
+    bool turnOffStream = false;
+
     _mainCamera.startImageStream((image) {
       FirebaseVisionImageMetadata metadata;
       //metadata tag for the for image format.
@@ -79,12 +79,16 @@ class _ScannerState extends State<Scanner> {
           .detectInImage(visionImage)
           .then((barcodes) {
         for (Barcode barcode in barcodes) {
-          Future.delayed(Duration(seconds: 2), () {
-            print("Running Future Delayed");
-            if (_scannedUids.add(barcode.rawValue)) {
-              pushToDB(barcode.rawValue);
+          if (!_scannedUids.contains(barcode.rawValue)) {
+            if (turnOffStream) {
+              print("Stream is turned off");
+            } else {
+              turnOffStream = true;
+              _scannedUids.add(barcode.rawValue);
+              pushToDB(barcode.rawValue)
+                  .then((onValue) => turnOffStream = false);
             }
-          });
+          }
         }
       }).catchError((error) {
         if (error.runtimeType == CameraException) {
@@ -175,7 +179,10 @@ class _ScannerState extends State<Scanner> {
   }
 
   void dispose() {
-    _mainCamera.dispose();
+    if (_mainCamera != null) {
+      _mainCamera.dispose();
+    }
+
     super.dispose();
   }
 
@@ -185,14 +192,14 @@ class _ScannerState extends State<Scanner> {
     isManualEnter = container.isManualEnter;
     //check first whether camera is init
     if (_isCameraInitalized) {
-        //check whether camera is already is streaming images
-        if (!_mainCamera.value.isStreamingImages) {
-          runStream();
-        }
-        //if there is an error, then stop the stream
-        else if (StateContainer.of(context).isThereConnectionError) {
-          _mainCamera.stopImageStream();
-        }
+      //check whether camera is already is streaming images
+      if (!_mainCamera.value.isStreamingImages) {
+        runStream();
+      }
+      //if there is an error, then stop the stream
+      else if (StateContainer.of(context).isThereConnectionError) {
+        _mainCamera.stopImageStream();
+      }
     }
 
     double screenHeight = MediaQuery.of(context).size.height;
@@ -296,12 +303,11 @@ class _ScannerState extends State<Scanner> {
     final container = StateContainer.of(context);
     double screenHeight = MediaQuery.of(context).size.height;
     double screenWidth = MediaQuery.of(context).size.width;
-    if(_isCameraInitalized)
-      {
-          _mainCamera.stopImageStream();
-      }
+    if (_isCameraInitalized) {
+      _mainCamera.stopImageStream();
+    }
     Navigator.pop(context);
-    Navigator.of(context).push(NoTransition(
-        builder: (context) => FinderScreen()));
+    Navigator.of(context)
+        .push(NoTransition(builder: (context) => FinderScreen()));
   }
 }
